Minor fixes related to Secrets Manager implementation
